Filters displaying on client side
