+++ client/app/bundles/SearchPage/components/searchPanel/quickFilterTab.jsx
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export default class QuickFilterTab extends Component {
+  render() {
+    return null;
+  }
+}
