+++ client/app/bundles/SearchPage/reducers/searchValue.js
@@ -1,7 +1,11 @@
 import { handleActions } from 'redux-actions';
+import { LOAD_STATE_FROM_PAGE } from '../actions/loadStateFromPage'
 
 export default handleActions({
   // [UPDATE_SEARCH_VALUE]: (state, action) => {
   //   return action.payload
   // }
+  [LOAD_STATE_FROM_PAGE]: (state, action) => {
+    return action.payload.searchValue;
+  }
 }, '')
